+++ src/ExpoYandexTransportModule.web.ts
@@ -1,11 +1,15 @@
 import { registerWebModule, NativeModule } from 'expo';
 
-import { Point, Route, RouteMode } from './ExpoYandexMapKit.types';
+import { DrivingRouteOptions, Point, Route, RouteMode } from './ExpoYandexMapKit.types';
 import { warnWebNotSupportedOnce } from './ExpoYandexMapKitModule.web';
 
 // Routing is not available on the web platform.
 class ExpoYandexTransportModule extends NativeModule {
-  async findRoutes(_points: Point[], _mode: RouteMode): Promise<Route[]> {
+  async findRoutes(
+    _points: Point[],
+    _mode: RouteMode,
+    _options?: DrivingRouteOptions | null
+  ): Promise<Route[]> {
     warnWebNotSupportedOnce();
     return [];
   }